requirements.txt
pytest==7.4.4
python-dotenv==0.21.0
requests==2.25.0
SimpleCache==0.1.2
Werkzeug==3.0.1
flask_caching==2.1.0
Flask~=3.0.1